Check if user id is duplicated before signup
